# core-lending-loan-servicing-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-servicing
  description: Loan Servicing Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-servicing
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/loan-servicing-cases/{caseId}/disbursements/{disbursementId}:
    get:
      tags:
        - LoanDisbursement
      summary: Get a disbursement by ID
      operationId: getDisbursementById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementDTO'
    put:
      tags:
        - LoanDisbursement
      summary: Update a disbursement record
      operationId: updateDisbursement
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDisbursementDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementDTO'
    delete:
      tags:
        - LoanDisbursement
      summary: Delete a disbursement record
      operationId: deleteDisbursement
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-servicing-cases/{caseId}/disbursements/{disbursementId}/internal-transactions:
    get:
      tags:
        - LoanDisbursementInternalTransaction
      summary: List/Search internal transactions for a disbursement
      operationId: findAllInternalTransactions
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.loanDisbursementId
          in: query
          description: Exact filter for loan disbursement id
          required: false
          schema:
            type: string
        - name: filters.transactionAmount
          in: query
          description: Exact filter for transaction amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[transactionAmount].from
          in: query
          description: Filter transaction amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[transactionAmount].to
          in: query
          description: Filter transaction amount to value
          required: false
          schema:
            type: string
        - name: filters.transactionReference
          in: query
          description: Exact filter for transaction reference
          required: false
          schema:
            type: string
        - name: filters.transactionDate
          in: query
          description: Exact filter for transaction date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].from
          in: query
          description: Filter transaction date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].to
          in: query
          description: Filter transaction date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseLoanDisbursementInternalTransactionDTO'
    post:
      tags:
        - LoanDisbursementInternalTransaction
      summary: Create a new internal transaction for a disbursement
      operationId: createInternalTransaction
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDisbursementInternalTransactionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementInternalTransactionDTO'
...
  /api/v1/loan-servicing-cases/{caseId}/disbursements/{disbursementId}/internal-transactions/{transactionId}:
    get:
      tags:
        - LoanDisbursementInternalTransaction
      summary: Get an internal transaction by ID
      operationId: getInternalTransactionById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoanDisbursementInternalTransactionDTO'
...
components:
  schemas:
...
    LoanDisbursementDTO:
      required:
        - disbursementAmount
        - disbursementDate
        - disbursementMethod
        - disbursementStatus
        - isFinalDisbursement
        - loanServicingCaseId
        - transactionId
      type: object
      properties:
        loanDisbursementId:
          type: string
          format: uuid
          readOnly: true
        loanServicingCaseId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        disbursementAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        disbursementDate:
          type: string
          format: date
        isFinalDisbursement:
          type: boolean
        disbursementMethod:
          type: string
          enum:
            - INTERNAL
            - EXTERNAL
        disbursementStatus:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - FAILED
            - REVERSED
        paymentProviderId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        distributorAgencyId:
          type: string
          format: uuid
        distributorAgentId:
          type: string
          format: uuid
        externalTransactionReference:
          maxLength: 255
          minLength: 0
          type: string
        note:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    LoanDisbursementInternalTransactionDTO:
      required:
        - destinationAccountId
        - loanDisbursementId
        - sourceAccountId
        - transactionAmount
        - transactionDate
      type: object
      properties:
        loanDisbursementInternalTransactionId:
          type: string
          format: uuid
          readOnly: true
        loanDisbursementId:
          type: string
          format: uuid
        sourceAccountId:
          type: string
          format: uuid
        destinationAccountId:
          type: string
          format: uuid
        transactionAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        transactionReference:
          maxLength: 255
          minLength: 0
          type: string
        transactionDate:
          type: string
          format: date-time
        note:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseLoanDisbursementInternalTransactionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/LoanDisbursementInternalTransactionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...